large integer support directly in Integer class
git-svn-id: svn://svn.code.sf.net/p/cgsuite/code/cgsuite/trunk@250 0138b134-6283-4d47-a013-9bc70bc83b06

+++ Language/src/org/cgsuite/lang/Domain.java
@@ -1418,7 +1418,7 @@ public class Domain
         {
             x = x.simplify();
             if (x instanceof CgsuiteInteger)
-                x = new CanonicalShortGame(((CgsuiteInteger) x).intValue());
+                x = new CanonicalShortGame((CgsuiteInteger) x);
             else if(x instanceof RationalNumber)
                 x = new CanonicalShortGame((RationalNumber) x);
             if (!(x instanceof CanonicalShortGame))
